renamed `events` package to `listener`
