+++ src/main/java/net/javadiscord/javabot/listener/StartupListener.java
@@ -1,4 +1,4 @@
-package net.javadiscord.javabot.events;
+package net.javadiscord.javabot.listener;
 
 
 import lombok.extern.slf4j.Slf4j;
